backend/tests/test_api.py
import pytest
from unittest.mock import MagicMock, patch
import sys

# Mocking heavy dependencies
mock_tf = MagicMock()
sys.modules["tensorflow"] = mock_tf
sys.modules["tensorflow.keras.models"] = mock_tf
sys.modules["tensorflow.keras.preprocessing.sequence"] = mock_tf
sys.modules["beanie"] = MagicMock()
sys.modules["motor"] = MagicMock()

from fastapi.testclient import TestClient
# We need to patch the init_db and ML loading in main to avoid side effects
with patch("app.main.init_db", return_value=None), \
     patch("app.main.load_ml_model", return_value=None):
    from app.main import app

client = TestClient(app)

def test_read_root():
    response = client.get("/")
    assert response.status_code == 200
    assert response.json() == {"message": "Welcome to Fake Job Detection API"}

def test_predict_endpoint_unauthorized():
    # Test predict endpoint without authentication if required
    # Looking at the routes, let's see if auth is needed
    response = client.post("/api/v1/predict/", json={"text": "test job"})
    # If auth is required, it should be 401 or 403
    # If not, it should be 200
    assert response.status_code in [200, 401, 403, 405] 
