backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor>=3.4.0
beanie==1.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
tensorflow-cpu==2.15.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
email-validator==2.1.0.post1
